+++ src/db/ColumnSchema.php
@@ -22,63 +22,63 @@ class ColumnSchema extends BaseObject
     /**
      * @var string name of this column (without quotes).
      */
-    public $name;
+    public string $name = '';
     /**
      * @var bool whether this column can be null.
      */
-    public $allowNull;
+    public bool $allowNull = false;
     /**
      * @var string abstract type of this column. Possible abstract types include:
      * char, string, text, boolean, smallint, integer, bigint, float, decimal, datetime,
      * timestamp, time, date, binary, and money.
      */
-    public $type;
+    public string $type = '';
     /**
      * @var string the PHP type of this column. Possible PHP types include:
      * `string`, `boolean`, `integer`, `double`, `array`.
      */
-    public $phpType;
+    public string $phpType = '';
     /**
      * @var string the DB type of this column. Possible DB types vary according to the type of DBMS.
      */
-    public $dbType;
+    public string $dbType = '';
     /**
      * @var mixed default value of this column
      */
-    public $defaultValue;
+    public mixed $defaultValue = null;
     /**
-     * @var array enumerable values. This is set only if the column is declared to be an enumerable type.
+     * @var array|null enumerable values. This is set only if the column is declared to be an enumerable type.
      */
-    public $enumValues;
+    public array|null $enumValues = null;
     /**
-     * @var int display size of the column.
+     * @var int|null display size of the column.
      */
-    public $size;
+    public int|null $size = null;
     /**
-     * @var int precision of the column data, if it is numeric.
+     * @var int|null precision of the column data, if it is numeric.
      */
-    public $precision;
+    public int|null $precision = null;
     /**
-     * @var int scale of the column data, if it is numeric.
+     * @var int|null scale of the column data, if it is numeric.
      */
-    public $scale;
+    public int|null $scale = null;
     /**
      * @var bool|null whether this column is a primary key
      */
-    public $isPrimaryKey;
+    public bool|null $isPrimaryKey = null;
     /**
      * @var bool whether this column is auto-incremental
      */
-    public $autoIncrement = false;
+    public bool $autoIncrement = false;
     /**
      * @var bool whether this column is unsigned. This is only meaningful
      * when [[type]] is `smallint`, `integer` or `bigint`.
      */
-    public $unsigned;
+    public bool $unsigned = false;
     /**
      * @var string comment of this column. Not all DBMS support this.
      */
-    public $comment;
+    public string $comment = '';
 
 
     /**
@@ -108,6 +108,35 @@ class ColumnSchema extends BaseObject
     public function defaultPhpTypecast($value)
     {
         return $this->phpTypecast($value);
+    }
+
+    /**
+     * Extracts size, precision, and scale from the dbType string.
+     *
+     * Parses parenthetical parameters from types like `varchar(255)` or `decimal(10,2)` and populates the corresponding
+     * column properties. Returns the base type name without parameters.
+     *
+     * @return string base type name (for example, `varchar` from `varchar(255)`).
+     */
+    public function extractSizeFromDbType(): string
+    {
+        if (preg_match('/^(\w+)(?:\(([^\)]+)\))?/', $this->dbType, $matches)) {
+            if (isset($matches[2]) && $matches[2] !== '') {
+                $values = explode(',', $matches[2]);
+
+                if (is_numeric($values[0])) {
+                    $this->size = $this->precision = (int) $values[0];
+
+                    if (isset($values[1])) {
+                        $this->scale = (int) $values[1];
+                    }
+                }
+            }
+
+            return $matches[1];
+        }
+
+        return $this->dbType;
     }
 
     /**